feat(Tracking): extract as IProcessable for two components
Extracted ArtificialVelocityApplierProcess to allow the component
to process in chosen moment.
Added convenient methods IncrementVelocity() and
IncrementAngularVelocity().
Use Vector3.Lerp() instead of Slerp() to calculate drag.
ArtificialVelocityApplier is derived from this new process
and maintains its current behaviour to process using
coroutine.

Extracted AverageVelocityEstimatorProcess to allow the
component to process in chosen moment.
AverageVelocityEstimator is derived from this new process
and maintains its current behaviour to process using
LateUpdate().
