+++ Runtime/Tracking/Velocity/ArtificialVelocityApplierProcess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc49ce751be9095429a51fc454ad7989
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
