+++ Runtime/Tracking/Velocity/AverageVelocityEstimatorProcess.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71d30dafbc9685b4ab2cd624b834d151
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
